Extracted rest api calls to client module
