+++ servicenow/tests/test_servicenow.py
@@ -17,7 +17,7 @@ from six import PY3
 from stackstate_checks.base import AgentIntegrationTestUtil, AgentCheck, TopologyInstance
 from stackstate_checks.base.errors import CheckException
 from stackstate_checks.base.stubs import topology, aggregator, telemetry
-from stackstate_checks.servicenow import ServicenowCheck, InstanceInfo, State
+from stackstate_checks.servicenow import ServiceNowCheck, InstanceInfo, State
 
 
 def mock_collect_process(*args):
@@ -293,7 +293,7 @@ class TestServicenow(unittest.TestCase):
         """
         Initialize and patch the check, i.e.
         """
-        self.check = ServicenowCheck('servicenow', {}, {}, [self.instance])
+        self.check = ServiceNowCheck('servicenow', {}, {}, [self.instance])
         topology.reset()
         aggregator.reset()
         telemetry.reset()
@@ -573,7 +573,7 @@ class TestServicenow(unittest.TestCase):
             }
         ]
         for test in tests:
-            check = ServicenowCheck('servicenow', {}, {}, [test['instance']])
+            check = ServiceNowCheck('servicenow', {}, {}, [test['instance']])
             result = json.loads(check.run())
             self.assertEqual(test['error'], result[0]['message'])
 
@@ -678,7 +678,7 @@ class TestServicenow(unittest.TestCase):
         Test max batch size value
         """
         instance = {'user': 'name', 'password': 'secret', 'url': "https://website.com", 'batch_size': 20000}
-        check = ServicenowCheck('servicenow', {}, {}, [instance])
+        check = ServiceNowCheck('servicenow', {}, {}, [instance])
         result = json.loads(check.run())
         self.assertEqual('{"batch_size": ["Int value should be less than or equal to 10000."]}', result[0]['message'])
 
@@ -688,7 +688,7 @@ class TestServicenow(unittest.TestCase):
         Test timeout exception exception gets critical service check
         """
         mock_request_get.side_effect = requests.exceptions.Timeout
-        check = ServicenowCheck('servicenow', {}, {}, [mock_instance])
+        check = ServiceNowCheck('servicenow', {}, {}, [mock_instance])
         check.run()
         service_checks = aggregator.service_checks(self.check.SERVICE_CHECK_NAME)
         self.assertEqual(1, len(service_checks))
